Revert "feat: visualize overlapping segments in cornerstone (#2185)"
This reverts commit 6b2cc499dd499d2b6780af8dca98116aad944af9.
