+++ extensions/dicom-segmentation/src/loadSegmentation.js
@@ -1,62 +1,92 @@
+import OHIF from '@ohif/core';
 import dcmjs from 'dcmjs';
 import cornerstone from 'cornerstone-core';
 import cornerstoneTools from 'cornerstone-tools';
 
+const { DicomLoaderService } = OHIF.utils;
+
 export default async function loadSegmentation(
-  imageIds,
   segDisplaySet,
-  labelmapBuffer,
-  segMetadata,
-  segmentsOnFrame,
-  labelmapSegments
+  referencedDisplaySet,
+  studies
 ) {
-  const { setters } = cornerstoneTools.getModule('segmentation');
+  const { StudyInstanceUID } = referencedDisplaySet;
 
-  // TODO: Could define a color LUT based on colors in the SEG.
-  const labelmapIndex = _getNextLabelmapIndex(imageIds[0]);
-  const colorLUTIndex = _makeColorLUTAndGetIndex(segMetadata);
+  // Set here is loading is asynchronous.
+  // If this function throws its set back to false.
+  segDisplaySet.isLoaded = true;
 
-  setters.labelmap3DByFirstImageId(
-    imageIds[0],
-    labelmapBuffer,
-    labelmapIndex,
-    segMetadata,
-    imageIds.length,
-    segmentsOnFrame,
-    colorLUTIndex
+  const segArrayBuffer = await DicomLoaderService.findDicomDataPromise(
+    segDisplaySet,
+    studies
   );
 
-  if (!segDisplaySet.labelmapSegments) {
-    segDisplaySet.labelmapSegments = {};
-  }
+  const dicomData = dcmjs.data.DicomMessage.readFile(segArrayBuffer);
+  const dataset = dcmjs.data.DicomMetaDictionary.naturalizeDataset(
+    dicomData.dict
+  );
 
-  /**
-   * Cache each labelmap segments.
-   * This data is used to determine the active label map when a given segment is activated/clicked.
-   */
-  segDisplaySet.labelmapSegments[labelmapIndex] = labelmapSegments.length ? Array.from(new Set(labelmapSegments.filter(a => !!a).reduce((a, b) => a.concat(b)))) : [];
-  segDisplaySet.labelmapIndex = labelmapIndex;
+  dataset._meta = dcmjs.data.DicomMetaDictionary.namifyDataset(dicomData.meta);
 
-  /*
-   * TODO: Improve the way we notify parts of the app that depends on segs to be loaded.
-   *
-   * Currently we are using a non-ideal implementation through a custom event to notify the segmentation panel
-   * or other components that could rely on loaded segmentations that
-   * the segments were loaded so that e.g. when the user opens the panel
-   * before the segments are fully loaded, the panel can subscribe to this custom event
-   * and update itself with the new segments.
-   *
-   * This limitation is due to the fact that the cs segmentation module is an object (which will be
-   * updated after the segments are loaded) that React its not aware of its changes
-   * because the module object its not passed in to the panel component as prop but accessed externally.
-   *
-   * Improving this event approach to something reactive that can be tracked inside the react lifecycle,
-   * allows us to easily watch the module or the segmentations loading process in any other component
-   * without subscribing to external events.
-   */
-  console.log('Segmentation loaded.');
-  const event = new CustomEvent('extensiondicomsegmentationsegloaded');
-  document.dispatchEvent(event);
+  const imageIds = _getImageIdsForDisplaySet(
+    studies,
+    StudyInstanceUID,
+    referencedDisplaySet.SeriesInstanceUID
+  );
+
+  return new Promise((resolve, reject) => {
+    let results;
+
+    try {
+      results = _parseSeg(segArrayBuffer, imageIds);
+    } catch (error) {
+      segDisplaySet.isLoaded = false;
+      segDisplaySet.loadError = true;
+      reject(error);
+    }
+
+    const { labelmapBuffer, segMetadata, segmentsOnFrame } = results;
+    const { setters } = cornerstoneTools.getModule('segmentation');
+
+    // TODO: Could define a color LUT based on colors in the SEG.
+    const labelmapIndex = _getNextLabelmapIndex(imageIds[0]);
+    const colorLUTIndex = _makeColorLUTAndGetIndex(segMetadata);
+
+    setters.labelmap3DByFirstImageId(
+      imageIds[0],
+      labelmapBuffer,
+      labelmapIndex,
+      segMetadata,
+      imageIds.length,
+      segmentsOnFrame,
+      colorLUTIndex
+    );
+
+    segDisplaySet.labelmapIndex = labelmapIndex;
+
+    /*
+     * TODO: Improve the way we notify parts of the app that depends on segs to be loaded.
+     *
+     * Currently we are using a non-ideal implementation through a custom event to notify the segmentation panel
+     * or other components that could rely on loaded segmentations that
+     * the segments were loaded so that e.g. when the user opens the panel
+     * before the segments are fully loaded, the panel can subscribe to this custom event
+     * and update itself with the new segments.
+     *
+     * This limitation is due to the fact that the cs segmentation module is an object (which will be
+     * updated after the segments are loaded) that React its not aware of its changes
+     * because the module object its not passed in to the panel component as prop but accessed externally.
+     *
+     * Improving this event approach to something reactive that can be tracked inside the react lifecycle,
+     * allows us to easily watch the module or the segmentations loading process in any other component
+     * without subscribing to external events.
+     */
+    console.log('Segmentation loaded.');
+    const event = new CustomEvent('extensiondicomsegmentationsegloaded');
+    document.dispatchEvent(event);
+
+    resolve(labelmapIndex);
+  });
 }
 
 function _getNextLabelmapIndex(firstImageId) {
@@ -141,4 +171,39 @@ function _getNextColorLUTIndex() {
   }
 
   return colorLUTIndex;
+}
+
+function _parseSeg(arrayBuffer, imageIds) {
+  return dcmjs.adapters.Cornerstone.Segmentation.generateToolState(
+    imageIds,
+    arrayBuffer,
+    cornerstone.metaData
+  );
+}
+
+function _getImageIdsForDisplaySet(
+  studies,
+  StudyInstanceUID,
+  SeriesInstanceUID
+) {
+  const study = studies.find(
+    study => study.StudyInstanceUID === StudyInstanceUID
+  );
+
+  const displaySets = study.displaySets.filter(displaySet => {
+    return displaySet.SeriesInstanceUID === SeriesInstanceUID;
+  });
+
+  if (displaySets.length > 1) {
+    console.warn(
+      'More than one display set with the same SeriesInstanceUID. This is not supported yet...'
+    );
+    // TODO -> We could make check the instance list and see if any match?
+    // Do we split the segmentation into two cornerstoneTools segmentations if there are images in both series?
+    // ^ Will that even happen?
+  }
+
+  const referencedDisplaySet = displaySets[0];
+
+  return referencedDisplaySet.images.map(image => image.getImageId());
 }
