+++ extensions/dicom-tag-browser/package.json
@@ -28,7 +28,7 @@
   },
   "peerDependencies": {
     "@ohif/core": "^2.6.0",
-    "dcmjs": "0.17.0",
+    "dcmjs": "0.16.5",
     "react": "^16.8.6"
   },
   "dependencies": {
