+++ platform/ui/src/contextProviders/SnackbarProvider.js
@@ -37,10 +37,6 @@ const SnackbarProvider = ({ children, service }) => {
         return null;
       }
 
-      if (options.type === 'error') {
-        console.error(options.error);
-      }
-
       const newItem = {
         ...DEFAULT_OPTIONS,
         ...options,
